A
-added a new Element in the beginning of the ModuleActivity called "Last
used Module" to test the SharedPreferences-function wich is now in use
in the ModuleSettingsScreen.
-changed the ModuleSettingsScreen. It now doesnt show the
xml/settings.xml anymore. Instead it uses
layout/activity_module_settings.xml again
-Made the OnCreate()-method in ModuleActivity more clear
MOST IMPORTANT CHANGE:
-after clicking on Add module you will see a simple settings screen. The
Info you type in there will be saved permanently and will be available
after a restart of the app and can be used to save modules