removed old stuff; added author/descript-comments
I have just updates all files but GameActivity.java because Leon is
working on it.
I have removed developer comments like: "this is shit, have to make it
better" and deleted some old xml files like the module.xml we tried to
use in the beginning.
In addition I added a comment before each class which determines the
author and contains a short description of this class.